# Backend/Frontend/openapi.raw.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Username or email already in use"
          }
        },
        "summary": "Create a new user account and issue a token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials, or account disabled"
          },
          "429": {
            "description": "Too many login attempts, try again later"
          }
        },
        "summary": "Authenticate with username/password and issue a token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired or reused refresh token"
          }
        },
        "summary": "Rotate a refresh token and issue a new token pair",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Always returned, regardless of whether the email exists or the account is active.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "If the account exists, a reset link has been sent."
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Request a password reset email. Always responds 202 with the same body, whether or not the address belongs to an account (anti-enumeration, P6 contract D13).",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Password reset"
          },
          "400": {
            "description": "Invalid or expired token — returned identically for every failure mode (malformed token, unknown/used/expired token, wrong secret, disabled account; P6 contract D12)."
          }
        },
        "summary": "Consume a password reset token: sets the new password and revokes every active refresh token for the account.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Refresh token revoked (idempotent: also returned for an already-invalid token)"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Revoke a refresh token",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get the currently authenticated user",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/tickets/{id}/attachments": {
      "post": {
        "operationId": "AttachmentsController_upload",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttachmentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "No file provided under \"file\""
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not the owner, the assignee, nor an admin (OwnershipGuard)"
          },
          "404": {
            "description": "Ticket not found"
          },
          "413": {
            "description": "File size exceeds UPLOAD_MAX_SIZE_BYTES"
          },
          "415": {
            "description": "MIME type not in UPLOAD_ALLOWED_MIME_TYPES"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upload an attachment to a ticket",
        "tags": [
          "attachments"
        ]
      },
      "get": {
        "operationId": "AttachmentsController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AttachmentResponseDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not the owner, the assignee, nor an admin (OwnershipGuard)"
          },
          "404": {
            "description": "Ticket not found"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List the attachments of a ticket",
        "tags": [
          "attachments"
        ]
      }
    },
    "/api/tickets/{id}/attachments/{attId}": {
      "delete": {
        "operationId": "AttachmentsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Attachment soft deleted"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "403": {
            "description": "Not the owner, the assignee, nor an admin (OwnershipGuard); or the caller can see the ticket but is neither the attachment's author nor an admin"
          },
          "404": {
            "description": "Ticket not found, or the attachment does not exist / does not belong to this ticket"
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Soft delete an attachment (its author, or an admin, only)",
        "tags": [
          "attachments"
        ]
      }
    },
...
  },
  "info": {
    "title": "Ticket Checker API",
    "description": "",
    "version": "1.0",
    "contact": {

    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UserRole": {
        "type": "string",
        "enum": [
          "ADMIN",
          "TECHNICIAN",
          "CLIENT"
        ],
        "description": "ADMIN or CLIENT. TECHNICIAN is rejected with 400 — use POST /technicians, which creates the account and its technician profile in one transaction (D1)."
      },
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "allOf": [
              {
                "$ref": "#/components/schemas/UserRole"
              }
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "firstName",
          "lastName",
          "phone",
          "role",
          "isActive",
          "createdAt"
        ]
      },
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "jdoe"
          },
          "email": {
            "type": "string",
            "example": "jdoe@example.com"
          },
          "password": {
            "type": "string",
            "example": "Str0ngP@ssw0rd",
            "minLength": 10,
            "maxLength": 72
          },
          "firstName": {
            "type": "string",
            "example": "Jane"
          },
          "lastName": {
            "type": "string",
            "example": "Doe"
          },
          "phone": {
            "type": "string",
            "example": "+1 555 123 4567"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ]
      },
      "AuthResponseDto": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIn0.abc123"
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIn0.def456"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponseDto"
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "identifier": {
            "type": "string",
            "example": "jdoe",
            "description": "Username or email"
          },
          "password": {
            "type": "string",
            "example": "Str0ngP@ssw0rd"
          }
        },
        "required": [
          "identifier",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxIn0.abc123"
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "jdoe@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6.k3f9...base64url-secret"
          },
          "newPassword": {
            "type": "string",
            "example": "Str0ngP@ssw0rd",
            "minLength": 10,
            "maxLength": 72
          }
        },
        "required": [
          "token",
          "newPassword"
        ]
      },
...
      "AttachmentResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "originalName": {
            "type": "string"
          },
          "mimeType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "downloadUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "originalName",
          "mimeType",
          "sizeBytes",
          "createdAt",
          "downloadUrl"
        ]
      },
...
    }
  }
}